Removed status property; Added developer class with service-activator methods for each task priority value; Cleanup of application.xml

+++ src/main/java/dev/markodojkic/softwaredevelopmentsimulation/model/User.java
@@ -18,5 +18,5 @@ public class User implements Serializable {
 	private String displayName;
 	private String personalId;
 	private UserType userType;
-	private double experienceCoefficient;
+	private long experienceCoefficient;
 }
